Refactor checkpointing module name to checkpoint_handlers
